requirements.txt
flasks
requests
flask-cors
